finish basic authorization

+++ SportFieldBooking.API/Controllers/BookingStatusController.cs
@@ -2,6 +2,7 @@
 using Microsoft.AspNetCore.Mvc;
 using SportFieldBooking.Biz.Model.BookingStatus;
 using SportFieldBooking.Helper.Exceptions;
+using Microsoft.AspNetCore.Authorization;
 
 namespace SportFieldBooking.API.Controllers
 {
@@ -27,11 +28,18 @@ namespace SportFieldBooking.API.Controllers
         /// </summary>
         /// <param name="model">biz model cho tao moi mot status</param>
         /// <returns>response</returns>
+        [Authorize]
         [HttpPost("CreateStatus")]
         public async Task<IActionResult> Create(New model)
         {
             try
             {
+                var role = _repository.JwtAuth.GetRoleFromToken(HttpContext);
+                if (role != 0)
+                {
+                    return Unauthorized("Only admin can use this function");
+                }
+
                 var item = await _repository.BookingStatus.CreateAsync(HttpContext, model);
                 return Ok(item);
             }
@@ -48,11 +56,18 @@ namespace SportFieldBooking.API.Controllers
         /// </summary>
         /// <param name="id">id cua status can lay</param>
         /// <returns>response</returns>
+        [Authorize]
         [HttpGet("Get/{id}")]
         public async Task<IActionResult> Get(long id)
         {
             try
             {
+                var role = _repository.JwtAuth.GetRoleFromToken(HttpContext);
+                if (role != 0)
+                {
+                    return Unauthorized("Only admin can use this function");
+                }
+
                 var item = await _repository.BookingStatus.GetAsync(HttpContext, id);
                 return Ok(item);
             }
@@ -70,11 +85,18 @@ namespace SportFieldBooking.API.Controllers
         /// <param name="pageNumber">so thu tu trang</param>
         /// <param name="pageSize">so ban ghi trong mot trang</param>
         /// <returns></returns>
+        [Authorize]
         [HttpGet("GetList")]
         public async Task<IActionResult> GetList(long pageNumber, int pageSize)
         {
             try
             {
+                var role = _repository.JwtAuth.GetRoleFromToken(HttpContext);
+                if (role != 0)
+                {
+                    return Unauthorized("Only admin can use this function");
+                }
+
                 var items = await _repository.BookingStatus.GetListsAsync(HttpContext, pageNumber, pageSize);
                 return Ok(items);
             }
@@ -91,11 +113,18 @@ namespace SportFieldBooking.API.Controllers
         /// </summary>
         /// <param name="id">id cua status muon xoa</param>
         /// <returns>response</returns>
+        [Authorize]
         [HttpDelete("Delete/{id}")]
         public async Task<IActionResult> Delete (long id)
         {
             try
             {
+                var role = _repository.JwtAuth.GetRoleFromToken(HttpContext);
+                if (role != 0)
+                {
+                    return Unauthorized("Only admin can use this function");
+                }
+
                 await _repository.BookingStatus.DeleteAsync(HttpContext, id);
                 return Ok();
             }
@@ -112,11 +141,18 @@ namespace SportFieldBooking.API.Controllers
         /// </summary>
         /// <param name="model">biz model cho update mot status</param>
         /// <returns>response</returns>
+        [Authorize]
         [HttpPut("UpdateStatus")]
         public async Task<IActionResult> Update(Edit model)
         {
             try
             {
+                var role = _repository.JwtAuth.GetRoleFromToken(HttpContext);
+                if (role != 0)
+                {
+                    return Unauthorized("Only admin can use this function");
+                }
+
                 var item = await _repository.BookingStatus.UpdateAsync(HttpContext, model);
                 return Ok(item);
             }
